requirements.txt
python-telegram-bot[webhooks]==20.3
werkzeug==2.3.7
quart==0.18.4
requests==2.32.3
python-dotenv==1.0.0
requests
pytz
asyncpg
loguru
regex